Write test for multi D3D instance bug
This happens on Nvidia if the window is resized. In that case the task manager
shows an increase for "dedicated GPU memory" in the Details tab for the process.
This memory seems to be a bug in how the value is calculated and not an actual
memory leak because the value can far exceed the available GPU+CPU memory.